Rename `panes_component` module to simply `panes`
