+++ src/platform/macos/panes.rs
